Benchmarking: run on non-Linux *nix systems

+++ waveform_benchmark/ioperf.py
@@ -1,3 +1,4 @@
+import math
 import os
 import re
 import resource
@@ -46,17 +47,25 @@ class PerformanceCounter:
             with _nocache_lock:
                 _nocache_enabled.add(self)
 
+        if self._clear_cache and not _nocache_supported:
+            self.n_bytes_read = math.nan
+
         # Run strace to track system calls.
-        self._strace = subprocess.Popen(['strace', '-c', '-f',
-                                         '-U', 'name,calls',
-                                         '-e', 'lseek,read',
-                                         '-p', str(os.getpid())],
-                                        stdout=subprocess.PIPE,
-                                        stderr=subprocess.STDOUT)
-        # Wait for strace to attach to the current process.  (Subtract
-        # one from n_read_calls because this counts as one.)
-        self._strace.stdout.readline()
-        self.n_read_calls -= 1
+        try:
+            self._strace = subprocess.Popen(['strace', '-c', '-f',
+                                             '-U', 'name,calls',
+                                             '-e', 'lseek,read',
+                                             '-p', str(os.getpid())],
+                                            stdout=subprocess.PIPE,
+                                            stderr=subprocess.STDOUT)
+            # Wait for strace to attach to the current process.  (Subtract
+            # one from n_read_calls because this counts as one.)
+            self._strace.stdout.readline()
+            self.n_read_calls -= 1
+        except FileNotFoundError:
+            self._strace = None
+            self.n_read_calls = math.nan
+            self.n_seek_calls = math.nan
 
         # Measure past resource usage for this process and any children.
         self._rusage_self = resource.getrusage(resource.RUSAGE_SELF)
@@ -73,14 +82,15 @@ class PerformanceCounter:
             _nocache_enabled.discard(self)
 
         # Stop the strace process and read its output.
-        self._strace.send_signal(signal.SIGINT)
-        strace_data, _ = self._strace.communicate()
-        for m in re.finditer(rb'^\s*(\w+)\s+(\d+)\s*$',
-                             strace_data, re.MULTILINE):
-            if m.group(1) == b'read':
-                self.n_read_calls += int(m.group(2))
-            elif m.group(1) == b'lseek':
-                self.n_seek_calls += int(m.group(2))
+        if self._strace is not None:
+            self._strace.send_signal(signal.SIGINT)
+            strace_data, _ = self._strace.communicate()
+            for m in re.finditer(rb'^\s*(\w+)\s+(\d+)\s*$',
+                                 strace_data, re.MULTILINE):
+                if m.group(1) == b'read':
+                    self.n_read_calls += int(m.group(2))
+                elif m.group(1) == b'lseek':
+                    self.n_seek_calls += int(m.group(2))
 
         # Calculate number of bytes read.  ru_inblock is always
         # measured in 512-byte blocks regardless of I/O block size.
@@ -100,25 +110,29 @@ class PerformanceCounter:
                              - self._rusage_children.ru_stime)
 
 
-def _nocache_hook(event, args):
-    global _nocache_loop
-    if event == 'open' and not _nocache_loop:
-        with _nocache_lock:
-            if _nocache_enabled and isinstance(args[0], (str, bytes)):
-                _nocache_loop = True
-                fd = None
-                try:
-                    fd = os.open(args[0], os.O_RDONLY | os.O_CLOEXEC)
-                    os.posix_fadvise(fd, 0, 0, os.POSIX_FADV_DONTNEED)
-                except OSError:
-                    pass
-                finally:
-                    if fd is not None:
-                        os.close(fd)
-                    _nocache_loop = False
-
-
 _nocache_lock = threading.Lock()
+_nocache_supported = False
 _nocache_enabled = set()
 _nocache_loop = False
-sys.addaudithook(_nocache_hook)
+
+if hasattr(os, 'posix_fadvise') and hasattr(os, 'POSIX_FADV_DONTNEED'):
+    _nocache_supported = True
+
+    def _nocache_hook(event, args):
+        global _nocache_loop
+        if event == 'open' and not _nocache_loop:
+            with _nocache_lock:
+                if _nocache_enabled and isinstance(args[0], (str, bytes)):
+                    _nocache_loop = True
+                    fd = None
+                    try:
+                        fd = os.open(args[0], os.O_RDONLY | os.O_CLOEXEC)
+                        os.posix_fadvise(fd, 0, 0, os.POSIX_FADV_DONTNEED)
+                    except OSError:
+                        pass
+                    finally:
+                        if fd is not None:
+                            os.close(fd)
+                        _nocache_loop = False
+
+    sys.addaudithook(_nocache_hook)
